# Conflicts:
#   src/ModelPackage/LoadFile.java
#   src/ModelPackage/SaveFile.java

+++ src/ModelPackage/IGizmo.java
@@ -15,7 +15,7 @@ public interface IGizmo {
 
     int getEndyPosition();
 
-    double getRotation();
+    int getRotation();
 
     Color getColour();
